+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     * Added `getFlexFeatures()` method to return all features that `FlexObject` or `FlexCollection` implements
     * Deprecated `FlexDirectory::update()` and `FlexDirectory::remove()`
     * Added `FlexStorage::getMetaData()` to get updated object meta information for listed keys
+    * Added `Language::getPageExtensions()` to get full list of supported page language extensions
 1. [](#improved)
     * Better support for Symfony local server `symfony server:start`
     * Make `Route` objects immutable